+++ tests/emr/test_emr_probes.py
@@ -49,7 +49,7 @@ def start_cluster():
 @mock_emr
 @pytest.mark.usefixtures("start_cluster")
 class TestEmrProbesMoto:
-    def setup(self, *args, **kwargs):
+    def setup_method(self, *args, **kwargs):
         client = boto3.client("emr", region_name="us-east-1")
         clusters = client.list_clusters()["Clusters"]
         self.cluster_id = clusters[0]["Id"]
@@ -82,7 +82,7 @@ class TestEmrProbesMoto:
 
 
 class TestEmrProbesMocked:
-    def setup(self):
+    def setup_method(self):
         self.cluster_id = "j-BC86D9AZOOHIA"
         self.group_id = "i-123456789ABCD"
         self.fleet_id = "i-987654321ZYWXW"